requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Interpretability
grad-cam>=1.4.8
captum>=0.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
albumentations>=1.3.1

# Metrics & Evaluation
scikit-image>=0.21.0
scipy>=1.11.0

# Development
jupyter>=1.0.0
ipykernel>=6.23.0
black>=23.3.0
flake8>=6.0.0
